refactor: move MockElasticHit to tests/__init__.py for shared usage
